new component structure implemented
